AiRXOS temp hack for OIDC compatibility
The GE OIDC user info endpoint does not comply strictly to the oidc
spec. The claims that are returned by GE openid are not the same as the
standard set that many client applications expect. For instance, this is
a sample user info response from GE

```
{
    "sub": "212462117",
    "country": "US",
    "employeetype": "Employee",
    "firstname": "Alex",
    "gehrbusinesssegment": "Aviation Avionics & Digital Systems",
    "mail": "alex.baluha@ge.com",
    "gessobusinessunit": "AVI ADS-AiRXOS",
    "gessouid": "9327087A-E714-1541-C0FF-002128B20D70",
    "title": "Lead DevOps Engineer",
    "georaclehrid": "212462117",
    "gevdsGroupIDmemberOf": [<omitted>],
    "lastname": "Baluha",
    "gessocompanyname": "General Electric Company",
    "gehrindustrygroup": "GE Aviation",
    "gessojobfunction": "Digital Technology",
    "street": "33-41 Farnsworth Street",
    "location": "Boston",
    "state": "MA",
    "SM_USER": "212462117"
}
```

The problem here, is that the standard set of claims expect `name`,
`email`, and `preferred_username` to exist. But GE doesn't include
those.

This commit modifies the plugin to use a custom string claim to use
`SM_USER` instead of `name`, and `mail` instead of `email`.

I didn't fix the tests that this change broke because I don't expect to
need to make any more changes in the future